sistema_padrones/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2026.1
django-simple-history==3.11.0
requests==2.33.1
djangorestframework-simplejwt==5.3.1
